Refactor AJAX calls in order to resolve #63
Created two new AJAX calls (customizer and after all import action).
The customizer AJAX request has the wp_customize set to 'on', so
it can use the customizer saving actions and that resolves #63 issue.

The after all import AJAX request is there to supersede the
customizer import and to enable users to hook after the whole
import process is over.